stored: separated VolumeCatalogInfo
